+++ workspaces/packages/pacwich/src/internal/core/index.ts
@@ -1,5 +1,6 @@
 export * from "./language";
 export * from "./error";
 export * from "./jsonc";
+export * from "./path";
 export * from "./process";
 export * from "./runtime";
