+++ workspaces/packages/pacwich/src/internal/core/path/index.ts
@@ -1,0 +1,1 @@
+export * from "./posixPath";
